+++ btools/building/fill/__init__.py
@@ -1,7 +1,7 @@
 import bpy
 
 from .fill_props import FillBars, FillPanel, FillLouver, FillGlassPanes
-from .fill_types import fill_bar, fill_panel, fill_louver, fill_glass_panes, FillUser
+from .fill_types import fill_face
 
 classes = (FillBars, FillPanel, FillLouver, FillGlassPanes)
